fix(pages/news/news): publică membru folosit în template
Repară eroare de compilare AoT.

"For a given component all its members (methods, properties) accessed by its template must be public in the ahead-of-time compilation scenario. This is due to the fact that a template is turned into a TS class. A generated class and a component are 2 separate classes now and you can't access private members cross-class." -- http://stackoverflow.com/a/39379451/2977191

Detalii:
- https://github.com/angular/angular/issues/11422
- http://blog.mgechev.com/2016/08/14/ahead-of-time-compilation-angular-offline-precompilation/#the-context-property